Refactored Driver and CGC to follow page object design pattern and seperated driver creation to a factory
