+++ selenium-tests/src/test/java/acceptance/CGCSteps.java
@@ -23,33 +23,39 @@ import cucumber.api.Transform;
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
-import driver.CGC;
+import driver.CGCPage;
+import driver.DriverFactory;
 import model.Node;
 import utils.Utils;
 
 public class CGCSteps {
 	public static String IFRAME_NAME = "iframe";
 	public static String CANVAS_XPATH = "//*[@id=\"canvas\"]";
+	
 	private HashMap<String, String> shortcuts;
 	private HashMap<String, Node> nodes;
-	private CGC cgc;
+	private CGCPage cgc;
+	private WebDriver driver;
+	
 	@Before
 	public void setUp() throws IOException
 	{
 		shortcuts = new HashMap<String, String>();
 		nodes = new HashMap<String, Node>();
+		driver = DriverFactory.createDriver();
+		cgc = new CGCPage(driver);
+		
 		shortcuts.put("Edge", "//*[@id=\"toolbar\"]/ul/li[2]/div");
-		cgc = new CGC();
 	}
 	@After
 	public void tearDown()
 	{
-		cgc.close();
+		driver.quit();
 	}
 	@Given("I navigate to the home page")
 	public void navigateTo() throws Throwable
 	{
-		cgc.load();
+		
 	}
 	@When("^.*do nothing$")
 	public void doNothing()
@@ -89,43 +95,12 @@ public class CGCSteps {
 	@Then("^*the screen should match '(.+)'$")
 	public void checkScreenshot(String path) throws Throwable
 	{
-	    cgc.assertContains(ImageIO.read(new File("src/test/resources/" + path)), 1);
+		cgc.assertContains("src/test/resources/" + path);
+	    //cgc.assertContains(ImageIO.read(new File("src/test/resources/" + path)), 1);
 	}
 	@Then("^.*there (?:is|are|should be) (.+) nodes?$")
 	public void checkNodes(String nodes) throws Throwable
 	{
-		cgc.assertContains(CGC.NODE_IMAGE, Utils.inNumerals(nodes));
+		cgc.assertContains(CGCPage.NODE_IMAGE, Utils.inNumerals(nodes));
 	}
-
-
- /**
-  * 	node1 = robot.createNode(33, 33)
-
-	#robot.save_screenshot('1_expected.png')
-	robot.assertScreenshot('1_expected.png')
-	node2 = robot.createNode(84, 84)
-
-	robot.edge(node1, node2)
-
-	robot.assertNode(node1)
-
-
-	robot.select(node1)
-	#node3 = robot.createNode(180, 180)
-
-
-	#robot.save_screenshot('2_expected.png')
-	robot.assertScreenshot('2_expected.png')
-
-	#robot.assertNode(node2)
-
-	robot.select(node1)
-	robot.select(node2, 10, 10)
-
-
-
-	#robot.save_screenshot('3_expected.png')
-	robot.assertScreenshot('3_expected.png')
-	return
-  */
 }
